requirements.txt
Flask==2.0.3
tensorflow==2.14.0  
numpy==1.21.6
pandas==1.3.5
requests==2.26.0
pywin32==227
matplotlib==3.4.3
scikit-learn==1.0.2
pillow==8.4.0
flask-cors==3.1.1
